fix: group cache
Laravel returns numeric strings when an int is saved to cache, which breaks a lot of the logic here. Fixed by removing cache everywhere except rss and middleware, and in those two spots, caching the whole group and not just the id of the group.

+++ app/Http/Controllers/RssController.php
@@ -132,7 +132,8 @@ class RssController extends Controller
     {
         $user = $request->user();
 
-        $disabledGroupId = cache()->rememberForever('group:disabled:id', fn () => Group::where('slug', '=', 'disabled')->soleValue('id'));
+        // Redis returns ints as numeric strings!
+        $disabledGroupId = (int) cache()->rememberForever('group:disabled:id', fn () => Group::where('slug', '=', 'disabled')->soleValue('id'));
 
         abort_if($user->group_id === $disabledGroupId, 404);
